Set up initial project files
With this commit, the project is initialized.
A README.md file outlining project goals is created.
An index.html file as well as a stylesheet and script file are
created and organized into relevant folders.
The index.html file is also given a very basic outline, including
linking it to the css and js files.