Align Display div to right and add fade to left
With this commit, the display div is styled to position text within it
in a paragraph that is aligned to the right.
Another div is placed atop, with a gradient, so that text appears to
be fading away to the left, if it should run out that far.
A variable is also now being used for the display color (and to practice
using CSS variables).