Connect number keys to js functions
With this commit, the number keys are connected to event listeners,
which listen for a click event and then call a numKeyPress function.
This numKeyPress function handles the input by adding each entry to an
inputArray and also displaying it to the calculator display.

To make this possible, num keys are given an additional class of num for
the querySelectorAll and also a data-key attribute matching their number
to make accessing this data from within the script easier.