Begin adding function to operation buttons
With this commit, functions for pressing of the operation keys and the
operate key are added.
These functions are not yet complete. They will still need all sorts of
functionality added to account for all sorts of circumstances that
a user might throw at them.

Some additional data attributes are added to the HTMl to help make this
function correctly.
Additionally, the clear and operate button have the class "function"
removed to help with query selection in the js. To compensate for this,
additional css is added to make sure the operate button is still
positioned correctly.