Clean up operation after operate function
Before this commit, when entering a new operation after an operation had
been tallied with the equals key, the calculator would automatically
perform the operation on the number itself, due to additional
functionality given to each operation key, designed to have a double
call on each key result in the opperation being performed on the number
itself.

With this commit, this functionality is adjusted by adding an additional
variable, afterOperate, which the operationKeyPress function looks for
as a condition when deciding what to do.
As a necessary thing, the clear and operationKeyPress functions now also
falsify this afterOperate boolean, to ensure correct functionality